+++ mceliese.h
@@ -1,6 +1,6 @@
 #include "linearcode.h"
 
-typedef std::tuple<size_t, size_t> permUnit;
+typedef std::tuple<uint64_t, uint64_t> permUnit;
 typedef std::vector<permUnit> permn;
 typedef std::tuple<matrix, LinearCode, permn> McEliesePrivate;
 typedef LinearCode McEliesePublic;
@@ -8,7 +8,7 @@ typedef LinearCode McEliesePublic;
 void applySwap(LinearCode& lc, const permUnit& swp);
 void applyPermn(LinearCode& lc, const permn& perm, const bool reverse=false);
 
-permn RandomPermutation(const size_t width, const size_t upto);
+permn RandomPermutation(const uint64_t width, const uint64_t upto);
 void PrintPermutation(const permn& ps);
 
 LinearCode PrivateToPublic(const McEliesePrivate& privKey);
